AI3DModelGenerator/requirements.txt
diffusers==0.28.0
transformers==4.39.0
accelerate==0.28.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
numpy>=1.24.0
trimesh>=4.0.0
open3d>=0.17.0
openai>=1.12.0
anthropic>=0.18.0
requests>=2.31.0
flask>=3.0.0
PyQt6>=6.6.0
pyopengl>=3.1.7
pyqtgraph>=0.13.0
numpy-stl>=3.0.0
plyfile>=1.0.0
objfile>=1.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
onnxruntime>=1.16.0
onnx>=1.15.0
pyinstaller>=6.3.0
aiohttp>=3.9.0
websockets>=12.0
scipy>=1.12.0
scikit-learn>=1.4.0
